Create laptrack_playground.py
Playground file to test LapTrack.

Main result is that you need to be careful with the 'alternative_cost_percentile' especially on small datasets, when gap closing seems weird

Further work is required to determine how to integrate other metrics (like color and intensity) in the process.
For now, it looks like the main way is to define a custom metric.