Detection Benchmark Infrastructure
Started the implementation of a benchmark for detection specifically.

It uses JSON files to store ground truth.

It also provides a temporary way to generate a JSON ground truth from a pipeline.
This file can be later edited by hand to correct the generation.

Ideally a small GUI will be added to do a real ground truth evaluation